condensed into different directories for slithering swimming and walking.
