+++ index.py
@@ -1,0 +1,41 @@
+from walking import Donkey, Llama, Goat, Sheep, Pig
+from slithering import Copperhead, RatSnake, CornSnake, BallPython, KingSnake
+from swimming import Mallard, Goldfish, Koi, Turtle, Frog
+
+# Creating instances of each class
+miss_fuzz = Llama(name="Miss Fuzz", species="Llama")
+billy = Goat(name="Billy", species="Goat")
+charlie = Donkey(name="Charlie", species="Donkey")
+wooly = Sheep(name="Wooly", species="Sheep")
+porky = Pig(name="Porky", species="Pig")
+
+slither = Copperhead(name="Slither", species="Copperhead")
+sneaky = RatSnake(name="Sneaky", species="Rat Snake")
+stripey = CornSnake(name="Stripey", species="Corn Snake")
+curly = BallPython(name="Curly", species="Ball Python")
+king = KingSnake(name="King", species="King Snake")
+
+quackers = Mallard(name="Quackers", species="Mallard")
+bubbles = Goldfish(name="Bubbles", species="Goldfish")
+splash = Koi(name="Splash", species="Koi")
+shellie = Turtle(name="Shellie", species="Turtle")
+ribbit = Frog(name="Ribbit", species="Frog")
+
+# Printing the name of each critter to verify
+print(miss_fuzz.name)
+print(billy.name)
+print(charlie.name)
+print(wooly.name)
+print(porky.name)
+
+print(slither.name)
+print(sneaky.name)
+print(stripey.name)
+print(curly.name)
+print(king.name)
+
+print(quackers.name)
+print(bubbles.name)
+print(splash.name)
+print(shellie.name)
+print(ribbit.name)
