+++ slithering/__init__.py
@@ -1,0 +1,5 @@
+from .copperhead import Copperhead
+from .ratsnake import RatSnake
+from .cornsnake import CornSnake
+from .ballpython import BallPython
+from .kingsnake import KingSnake
