+++ slithering/ratsnake.py
@@ -1,0 +1,8 @@
+from datetime import date
+
+class RatSnake:
+        def __init__(self, name, species):
+            self.name = name
+            self.species = species
+            self.date_added = date.today()
+            self.slithering = True
