+++ swimming/__init__.py
@@ -1,0 +1,5 @@
+from .mallard import Mallard
+from .goldfish import Goldfish
+from .koi import Koi
+from .turtle import Turtle
+from .frog import Frog
