+++ swimming/goldfish.py
@@ -1,0 +1,8 @@
+from datetime import date
+
+class Goldfish:
+    def __init__(self, name, species):
+        self.name = name
+        self.species = species
+        self.date_added = date.today()
+        self.swimming = True
