+++ walking/__init__.py
@@ -1,0 +1,5 @@
+from .donkey import Donkey
+from .llama import Llama
+from .goat import Goat
+from .sheep import Sheep
+from .pig import Pig
