+++ walking/goat.py
@@ -1,0 +1,8 @@
+from datetime import date
+
+class Goat:
+    def __init__(self, name, species):
+        self.name = name
+        self.species = species
+        self.date_added = date.today()
+        self.walking = True
